+++ src/github.com/getlantern/flashlight/flashlight.go
@@ -12,12 +12,10 @@ import (
 	"runtime/pprof"
 	"time"
 
-	"github.com/getlantern/fronted"
 	"github.com/getlantern/golog"
 
 	"github.com/getlantern/flashlight/client"
 	"github.com/getlantern/flashlight/config"
-	"github.com/getlantern/flashlight/server"
 	"github.com/getlantern/flashlight/statreporter"
 	"github.com/getlantern/flashlight/statserver"
 )
@@ -133,38 +131,6 @@ func runClientProxy(cfg *config.Config) {
 	err := client.ListenAndServe()
 	if err != nil {
 		log.Fatalf("Unable to run client proxy: %s", err)
-	}
-}
-
-// Runs the server-side proxy
-func runServerProxy(cfg *config.Config) {
-	useAllCores()
-
-	srv := &server.Server{
-		Addr:         cfg.Addr,
-		Host:         cfg.Server.AdvertisedHost,
-		ReadTimeout:  0, // don't timeout
-		WriteTimeout: 0,
-		CertContext: &fronted.CertContext{
-			PKFile:         config.InConfigDir("proxypk.pem"),
-			ServerCertFile: config.InConfigDir("servercert.pem"),
-		},
-	}
-
-	srv.Configure(cfg.Server)
-
-	// Continually poll for config updates and update server accordingly
-	go func() {
-		for {
-			cfg := <-configUpdates
-			configureStats(cfg, false)
-			srv.Configure(cfg.Server)
-		}
-	}()
-
-	err := srv.ListenAndServe()
-	if err != nil {
-		log.Fatalf("Unable to run server proxy: %s", err)
 	}
 }
