+++ src/github.com/getlantern/flashlight/server.go
@@ -1,0 +1,41 @@
+// +build darwin dragonfly freebsd !android,linux netbsd openbsd solaris
+
+package main
+
+import (
+	"github.com/getlantern/flashlight/config"
+	"github.com/getlantern/flashlight/server"
+	"github.com/getlantern/fronted"
+)
+
+// Runs the server-side proxy
+func runServerProxy(cfg *config.Config) {
+	useAllCores()
+
+	srv := &server.Server{
+		Addr:         cfg.Addr,
+		Host:         cfg.Server.AdvertisedHost,
+		ReadTimeout:  0, // don't timeout
+		WriteTimeout: 0,
+		CertContext: &fronted.CertContext{
+			PKFile:         config.InConfigDir("proxypk.pem"),
+			ServerCertFile: config.InConfigDir("servercert.pem"),
+		},
+	}
+
+	srv.Configure(cfg.Server)
+
+	// Continually poll for config updates and update server accordingly
+	go func() {
+		for {
+			cfg := <-configUpdates
+			configureStats(cfg, false)
+			srv.Configure(cfg.Server)
+		}
+	}()
+
+	err := srv.ListenAndServe()
+	if err != nil {
+		log.Fatalf("Unable to run server proxy: %s", err)
+	}
+}
